requirements.txt
torch==2.5.1
transformers==4.47.1
lightning==2.4.0
hydra-core==1.3.2
hydra-colorlog==1.2.0
torchmetrics==0.10.3
peft==0.14.0
datasets==3.1.0
numpy==2.2.5
tqdm==4.67.1
rootutils==1.0.7
scikit-learn==1.6.1
rich==13.9.4
accelerate==1.1.1
wandb==0.19.0
